build(firefox): check if profile.ini exists first

+++ firefox/install.sh
@@ -7,10 +7,10 @@ if ! command -v "$pkgname" >/dev/null; then
 	exit 0
 fi
 
-printf "%s: Installing user.js\n" "$0"
+if [ -e ~/.mozilla/firefox/profiles.ini ] && [ -d "$XDG_CONFIG_HOME"/firefox ]; then
+	printf "%s: Installing user.js\n" "$0"
 
-profile=$(grep -Po -m1 'Default=\K.*' <~/.mozilla/firefox/profiles.ini)
-if [ -d "$XDG_CONFIG_HOME"/firefox ]; then
+	profile=$(grep -Po -m1 'Default=\K.*' <~/.mozilla/firefox/profiles.ini)
 	for item in "$XDG_CONFIG_HOME"/firefox/*; do
 		if [ ! -L ~/.mozilla/firefox/"$profile"/"${item##*/}" ]; then
 			ln -vst ~/.mozilla/firefox/"$profile" "$item"
